requirements.txt
Flask
keras
numpy
Pillow
werkzeug
tensorflow
scikit-image
scipy==1.1.0
pandas
sklearn
